Fix bug #922247, PKCS12Decoder.database_import() method. Importing into a NSS database would sometimes fail or segfault.
PKCS12DecoderStart() creates a decoder context in which it caches
state. A pointer to SECItem with the PKCS12 password is passed to
PKCS12DecoderStart(), but instead of copying the SECItem into the
context it copied just the pointer to the SECITem which had been
allocated on the stack. The import operation occurs after the stack
frame with the SECItem had been released thus the decoder context
referenced invalid stack memory which may have been reused. The fix is
to allocate the SECItem in the PKCS12Decoder python object along with
the decoder context.