Unify subprocess calling & enhance subprocess error handling
Replace duplicated code blocks invoking subprocess.Popen with call to
new function rom_cmd(). Enhance the CmdError exception to also record
the command arguments and stdout. The CmdError message is now optional
and will be constructed from the input parameters. Remove 1 cmd_args
block that was accidently left in as the result of a cut-n-paste
error.