Strengthen test_pkcs12 by checking for expected values.
test_pkcs12 was a weak test, it essentially just tried to perform
pkcs12 import and export operations and assumed if there were no
exceptions the operation was successful. Not checking the expected
results was probably due to the perceived difficulty of getting the
expected data; keys are never exposed by NSS and pk12 files which
contain binary identical data are themselves not binary identical due
to the password based encryption using a different salt. Also the test
had a number of confusing variable names and operations that made the
test less than clear. The following changes were made:

* Only one key/cert is now used in the test (test_user), the use of
  the test_server key/cert was removed, this simplified the logic and
  naming without any significant loss in testing.

* All variables prefixed with 'read' were renamed to strip that
  prefix. The read prefix had been used because it was the pk12 file
  we were reading, but the same variables were being used in contexts
  other than reading the pk12 so it was confusing.

* The CmdError exception was enhanced to provide more information.

* The run_cmd function was enhanced to be more feature full and
  robust.

* Calls to run shell utilities were factored out into their own
  functions to improve readability, these included:
  get_cert_der_from_db(), delete_cert_from_db(), create_pk12(),
  list_pk12()

* The test_user cert in DER format is read from the database at the
  beginnig of each test to assure it's present (a necessary condition)
  and at the end of the test it's read again from the database to
  assure it's still present and unmodified.

* The exported pk12 file is now validated against the original pk12
  file. This is done by calling pk12util to produce an ASCII dump of
  the file and discarding the key portion which may not be identical
  due to the salt.